requirements.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
pydantic[email]==1.10.2
sqlmodel==0.0.8
pyjwt[crypto]==2.6.0
sendgrid==6.9.7
psycopg2-binary==2.9.5
gunicorn==20.1.0
loguru==0.6.0

